split shim up, make debugging configurable
